Move glob into core
