+++ src/core/BUILD
@@ -10,10 +10,13 @@ genrule(
 
 go_library(
     name = 'core',
-    srcs = glob(['*.go'], excludes = [
-        '*_test.go',
-        'version.go',
-    ]) + [':version'],
+    srcs = glob(
+        ['*.go'],
+        excludes = [
+            '*_test.go',
+            'version.go',
+        ],
+    ) + [':version'],
     visibility = ['PUBLIC'],
     deps = [
         '//third_party/go:gcfg',
@@ -96,6 +99,16 @@ go_test(
 go_test(
     name = 'lock_test',
     srcs = ['lock_test.go'],
+    deps = [
+        ':core',
+        '//third_party/go:testify',
+    ],
+)
+
+go_test(
+    name = 'glob_test',
+    srcs = ['glob_test.go'],
+    data = glob(['test_data/**/*.txt']),
     deps = [
         ':core',
         '//third_party/go:testify',
