+++ src/core/glob_test.go
@@ -1,20 +1,18 @@
 // Tests for our glob functions.
 
-package parse
+package core
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
-	"core"
 )
 
 func TestCanGlobFirstFile(t *testing.T) {
 	// If this fails then we probably failed to interpret /**/ properly,
 	// which can resolve to just / - ie. we glob test_data/**/*.txt,
 	// which should include test_data/test.txt
-	if !core.FileExists("src/parse/test_data/test.txt") {
+	if !FileExists("src/core/test_data/test.txt") {
 		t.Errorf("Can't load test_data/test.txt")
 	}
 }
@@ -23,20 +21,20 @@ func TestCanGlobSecondFile(t *testing.T) {
 	// If this fails then we haven't walked down enough subdirectories
 	// or something. Shouldn't really be hard - it's a sanity check really
 	// since it's similar to the third file but without a package boundary.
-	if !core.FileExists("src/parse/test_data/test_subfolder1/a.txt") {
+	if !FileExists("src/core/test_data/test_subfolder1/a.txt") {
 		t.Errorf("Can't load test_data/test_subfolder1/a.txt")
 	}
 }
 
 func TestCannotGlobThirdFile(t *testing.T) {
 	// This one we should not be able to glob because it's inside its own subpackage.
-	if core.FileExists("src/parse/test_data/test_subfolder2/b.txt") {
+	if FileExists("src/core/test_data/test_subfolder2/b.txt") {
 		t.Errorf("Incorrectly loaded test_data/test_subfolder2/b.txt; have globbed it through a package boundary")
 	}
 }
 
 func TestCanGlobFileAtRootWithDoubleStar(t *testing.T) {
-	files, err := glob("src/parse/test_data/test_subfolder1", "**/*.txt", false, []string{})
+	files, err := glob("src/core/test_data/test_subfolder1", "**/*.txt", false, []string{})
 	assert.NoError(t, err)
-	assert.Equal(t, []string{"src/parse/test_data/test_subfolder1/a.txt"}, files)
+	assert.Equal(t, []string{"src/core/test_data/test_subfolder1/a.txt"}, files)
 }
