+++ src/core/test_data/TEST_BUILD
@@ -1,0 +1,1 @@
+# Not a real BUILD file, used in interpreter_test.
