+++ src/core/test_data/test_subfolder2/BUILD
@@ -1,0 +1,2 @@
+# This BUILD file is here to test that rules from parent
+# directories can't glob its files.
